+++ tests/testthat/test-spatialInference.R
@@ -57,6 +57,7 @@ res <- spatialInference(spe, c("alpha", "Tc"),
                         sample_id = "patient_id",
                         image_id = "image_number", condition = "patient_stage",
                         ncores = 1,
+                        intensityAdjustment = FALSE,
                         algorithm = "bam", discrete = FALSE,
                         eps = NULL, delta = 0, family = stats::gaussian(link = "identity"),
                         AR1 = FALSE, weightTransform = FALSE, sandwich = FALSE
@@ -208,6 +209,19 @@ test_that("spatilInference can handle completely missing points",{
     algorithm = "bam"
   )
   expect_true(is.null(res$mdl) && is.null(res$designmat))
+})
+
+test_that("spatilInference works with intensity covariate adjustement",{
+  res <- spatialInference(spe, c("alpha", "Tc"),
+    fun = "Gcross", marks = "cell_type",
+    rSeq = seq(0, 50, length.out = 50), correction = "rs",
+    sample_id = "patient_id",
+    image_id = "image_number", condition = "patient_stage",
+    ncores = 1,
+    algorithm = "bam",
+    intensityAdjustment = TRUE
+  )
+  expect_true(!is.null(res$mdl))
 })